Moved deferred matcher/setter code to its own package.
