+++ onebusaway-gtfs-transformer/src/main/java/org/onebusaway/gtfs_transformer/deferred/DeferredValueMatcher.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.onebusaway.gtfs_transformer.impl;
+package org.onebusaway.gtfs_transformer.deferred;
 
 import org.apache.commons.beanutils.Converter;
 import org.onebusaway.gtfs.model.AgencyAndId;
