+++ onebusaway-gtfs-transformer/src/main/java/org/onebusaway/gtfs_transformer/factory/TransformFactory.java
@@ -60,14 +60,14 @@ import org.onebusaway.gtfs_transformer.collections.ServiceIdKey;
 import org.onebusaway.gtfs_transformer.collections.ServiceIdKeyMatch;
 import org.onebusaway.gtfs_transformer.collections.ShapeIdKey;
 import org.onebusaway.gtfs_transformer.collections.ShapeIdKeyMatch;
-import org.onebusaway.gtfs_transformer.impl.DeferredValueMatcher;
-import org.onebusaway.gtfs_transformer.impl.DeferredValueSetter;
-import org.onebusaway.gtfs_transformer.impl.EntitySchemaCache;
+import org.onebusaway.gtfs_transformer.deferred.DeferredValueMatcher;
+import org.onebusaway.gtfs_transformer.deferred.DeferredValueSetter;
+import org.onebusaway.gtfs_transformer.deferred.EntitySchemaCache;
+import org.onebusaway.gtfs_transformer.deferred.ValueSetter;
 import org.onebusaway.gtfs_transformer.impl.RemoveEntityUpdateStrategy;
 import org.onebusaway.gtfs_transformer.impl.ServiceIdTransformStrategyImpl;
 import org.onebusaway.gtfs_transformer.impl.SimpleModificationStrategy;
 import org.onebusaway.gtfs_transformer.impl.StringModificationStrategy;
-import org.onebusaway.gtfs_transformer.impl.ValueSetter;
 import org.onebusaway.gtfs_transformer.match.EntityMatch;
 import org.onebusaway.gtfs_transformer.match.EntityMatchCollection;
 import org.onebusaway.gtfs_transformer.match.PropertyAnyValueEntityMatch;
