+++ onebusaway-gtfs-transformer/src/main/java/org/onebusaway/gtfs_transformer/impl/SimpleModificationStrategy.java
@@ -20,6 +20,7 @@ import java.util.Map;
 import org.onebusaway.csv_entities.schema.BeanWrapper;
 import org.onebusaway.csv_entities.schema.BeanWrapperFactory;
 import org.onebusaway.gtfs.services.GtfsMutableRelationalDao;
+import org.onebusaway.gtfs_transformer.deferred.ValueSetter;
 import org.onebusaway.gtfs_transformer.services.EntityTransformStrategy;
 import org.onebusaway.gtfs_transformer.services.TransformContext;
